requirements.txt
google-ads>=23.0.0
python-dotenv>=1.0.0